move monotime
