+++ be/src/storage/data_dir.cpp
@@ -40,6 +40,7 @@
 #include <utility>
 
 #include "base/system/errno.h"
+#include "base/time/monotime.h"
 #include "base/utility/defer_op.h"
 #include "common/config.h"
 #include "fs/fs.h"
@@ -57,7 +58,6 @@
 #include "storage/tablet_updates.h"
 #include "storage/txn_manager.h"
 #include "storage/utils.h" // for check_dir_existed
-#include "util/monotime.h"
 #include "util/string_util.h"
 
 using strings::Substitute;
