+++ be/src/util/countdown_latch.h
@@ -37,9 +37,9 @@
 #include <condition_variable>
 #include <mutex>
 
+#include "base/time/monotime.h"
 #include "common/logging.h"
 #include "storage/olap_define.h"
-#include "util/monotime.h"
 
 namespace starrocks {
