+++ be/src/util/gc_helper.h
@@ -15,7 +15,7 @@
 #pragma once
 
 #include "base/gc/gc_helper_smoothstep.h"
-#include "util/monotime.h"
+#include "base/time/monotime.h"
 
 namespace starrocks {
