+++ be/test/runtime/batch_write/isomorphic_batch_write_test.cpp
@@ -18,12 +18,12 @@
 
 #include "base/testutil/assert.h"
 #include "base/testutil/sync_point.h"
+#include "base/time/monotime.h"
 #include "gen_cpp/FrontendService.h"
 #include "http/http_common.h"
 #include "runtime/stream_load/stream_load_context.h"
 #include "runtime/stream_load/time_bounded_stream_load_pipe.h"
 #include "util/bthreads/executor.h"
-#include "util/monotime.h"
 #include "util/threadpool.h"
 
 namespace starrocks {
