+++ be/test/runtime/fragment_mgr_test.cpp
@@ -38,11 +38,11 @@
 
 #include <utility>
 
+#include "base/time/monotime.h"
 #include "common/config.h"
 #include "exec/data_sink.h"
 #include "runtime/exec_env.h"
 #include "runtime/plan_fragment_executor.h"
-#include "util/monotime.h"
 
 namespace starrocks {
