+++ be/test/util/countdown_latch_test.cpp
@@ -21,8 +21,8 @@
 
 #include <functional>
 
+#include "base/time/monotime.h"
 #include "gutil/ref_counted.h"
-#include "util/monotime.h"
 #include "util/thread.h"
 #include "util/threadpool.h"
