+++ be/test/util/threadpool_test.cpp
@@ -39,6 +39,7 @@
 #include "base/concurrency/spinlock.h"
 #include "base/random/random.h"
 #include "base/testutil/assert.h"
+#include "base/time/monotime.h"
 #include "base/utility/scoped_cleanup.h"
 #include "common/logging.h"
 #include "common/status.h"
@@ -50,7 +51,6 @@
 #include "gutil/walltime.h"
 #include "util/countdown_latch.h"
 #include "util/metrics.h"
-#include "util/monotime.h"
 
 using std::atomic;
 using std::shared_ptr;
