+++ be/test/util/trace_test.cpp
@@ -33,13 +33,13 @@
 #include <vector>
 
 #include "base/concurrency/stopwatch.hpp"
+#include "base/time/monotime.h"
 #include "base/utility/scoped_cleanup.h"
 #include "gutil/macros.h"
 #include "gutil/port.h"
 #include "gutil/ref_counted.h"
 #include "gutil/walltime.h"
 #include "util/countdown_latch.h"
-#include "util/monotime.h"
 #include "util/thread.h"
 #include "util/trace_metrics.h"
